test(cli): move VCR support outside shipped sources
